core/requirements.txt
pymongo==2.8
Flask==0.10.1
mongokit-0.9.1.1